LayoutTests/imported/w3c/web-platform-tests/fetch/sec-metadata/report.tentative.https.sub-expected.txt

CONSOLE MESSAGE: Refused to load https://foo.bar/ because it does not appear in the style-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://foo.bar/ because it does not appear in the style-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://foo.bar/ because it does not appear in the style-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://foo.bar/ because it does not appear in the style-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://foo.bar/ because it does not appear in the style-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://foo.bar/ because it does not appear in the style-src directive of the Content Security Policy.

PASS Initialization. 
